+++ src/main/javascript/webpack/BuildUtils.js
@@ -1,5 +1,6 @@
 "use strict";
 const path = require('path');
+const fs = require('fs');
 
 function artifactName(projectName, baseName) {
   const nameParts = projectName.replace(/\+/, '').split(' ');
@@ -29,6 +30,32 @@ function extractVendors (packageJson)
   return vendorPackages;
 }
 
+function resolveBabelOptions(babelOptions)
+{
+  const resolvePluginOrPreset = function (module) {
+    if (typeof module === 'string') { return require.resolve(module); }
+    return module;
+  };
+
+  if (babelOptions.presets) {
+    babelOptions.presets = babelOptions.presets.map(resolvePluginOrPreset)
+  }
+
+  if (babelOptions.env) {
+    Object.keys(babelOptions.env).forEach(function(env) {
+      if (babelOptions.env[env].plugins) {
+        babelOptions.env[env].plugins = babelOptions.env[env].plugins.map(resolvePluginOrPreset);
+      }
+    });
+  }
+
+  if (babelOptions.plugins) {
+    babelOptions.plugins = babelOptions.plugins.map(resolvePluginOrPreset);
+  }
+
+  return babelOptions;
+}
+
 module.exports = {
   artifactName: function (projectDir) {
     "use strict";
@@ -39,10 +66,68 @@ module.exports = {
       return artifactName(projectName, baseName);
     }
   },
-  autoVendorPackages: function (projectDir) {
+
+  /**
+   * Scans the package.json manifest for dependencies and devdependencies and returns a list of package names that
+   * can be used by webpack to create the vendor bundle
+   *
+   * @param {String} projectDir
+   * @return {Array<String>}
+   */
+  autoVendorDependencies: function (projectDir) {
     "use strict";
     const packageJsonPath = path.resolve(projectDir, 'package.json');
     const packageJson = require(packageJsonPath);
-    return extractVendors(packageJson);
+
+    const exclusions = [
+      '@deskproapps/dpat'
+    ];
+
+    const extracted = extractVendors(packageJson);
+    return extracted.filter(function (packageName) {
+      return exclusions.indexOf(packageName) === -1;
+    });
+  },
+
+  /**
+   * Replaces plugin and presets short names with absolute paths
+   *
+   * @param {{}} options
+   * @param {{}} overrides
+   * @return {*}
+   */
+  resolveBabelOptions: function (options, overrides) {
+    let resolved = null;
+
+    if (typeof options === 'string') { //we've been give a path to a project dir or babelrc
+      let babelrcPath;
+      const stats = fs.statSync(options);
+      if (stats.isDirectory()) {
+        babelrcPath = path.join(options, '.babelrc');
+      } else if (stats.isFile()) {
+        babelrcPath = options;
+      }
+
+      if (babelrcPath) {
+        try {
+          const babelOptions = JSON.parse(fs.readFileSync(babelrcPath));
+          resolved = resolveBabelOptions(babelOptions);
+        } catch (e) {
+          console.log(e);
+        }
+      }
+
+    } else if (typeof options === 'object') {
+      resolved = resolveBabelOptions(options);
+    }
+
+    if (resolved) {
+      Object.keys(overrides).forEach(function (key) {
+        resolved[key] = overrides[key];
+      });
+      return resolved;
+    }
+
+    return null;
   }
 };